+++ srcs/transformations.c
@@ -6,7 +6,7 @@
 /*   By: tkeil <tkeil@student.42heilbronn.de>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 18:27:00 by tkeil             #+#    #+#             */
-/*   Updated: 2024/11/27 13:14:41 by tkeil            ###   ########.fr       */
+/*   Updated: 2024/11/28 01:40:57 by tkeil            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	ft_translate(t_point *point, float x, float y)
 {
-	point->x += (int)(x);
-	point->y += (int)(y);
+	point->x += x;
+	point->y += y;
 }
 
 void	ft_rotate(t_point *p, float w, void (*f)(t_point *, float))
@@ -23,21 +23,7 @@ void	ft_rotate(t_point *p, float w, void (*f)(t_point *, float))
 	f(p, w * M_PI / 180);
 }
 
-// void	ft_scale(t_wire *wire, t_point *point, float fac, int pad)
-// {
-// 	float	auto_scale;
-// 	float	f_x;
-// 	float	f_y;
-
-// 	f_x = (WIDTH - pad) / wire->max_w;
-// 	f_y = (HEIGHT - pad) / wire->height;
-// 	auto_scale = (f_x * (f_x <= f_y) + f_y * (f_y < f_x));
-// 	point->x = (int)round(point->x * auto_scale * fac);
-// 	point->y = (int)round(point->y * auto_scale * fac);
-// 	point->z = (int)round(point->z * auto_scale * fac);
-// }
-
-void	ft_scale(t_wire **wire, float fac, int pad)
+void	ft_scale(t_wire **wire, float fac, float pad)
 {
 	int		i;
 	int		j;
@@ -55,13 +41,11 @@ void	ft_scale(t_wire **wire, float fac, int pad)
 		j = 0;
 		while (j < (*wire)->widths[i])
 		{
-			p[i][j].x = (int)(p[i][j].x * (*wire)->auto_scale * fac);
-			p[i][j].y = (int)(p[i][j].y * (*wire)->auto_scale * fac);
-			p[i][j].z = (int)(p[i][j].z * (*wire)->auto_scale * fac);
+			p[i][j].x = p[i][j].x * (*wire)->auto_scale * fac;
+			p[i][j].y = p[i][j].y * (*wire)->auto_scale * fac;
+			p[i][j].z = p[i][j].z * (*wire)->auto_scale * fac;
 			j++;
 		}
 		i++;
 	}
-	// ft_set_max_height(*wire);
-	// ft_set_max_width(*wire);
 }
